% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camt.cor.func.R
\name{camt.fwer}
\alias{camt.fwer}
\title{Perform the covariate-adaptive family-wise error rate control}
\usage{
camt.fwer(
  pvals,
  pi0.var,
  data = data,
  pvals.cutoff = 1e-15,
  alpha = 0.1,
  k.fix = NULL,
  EM.paras = list(iterlim = 50, tol = 1e-05, k.init = NULL, pi0.init = NULL, nlm.iter =
    5),
  trace = FALSE,
  return.model.matrix = FALSE
)
}
\arguments{
\item{pvals}{a numeric vector of p-values.}

\item{pi0.var}{a formula, a vector, a data frame, or a matrix of covariate values for the prior null probability.}

\item{data}{a data frame containing the covariates, only used when pi0.var are classes of 'formula'.}

\item{pvals.cutoff}{a numeric value to replace p-values below that value, which is used to increase the stability of the algorithm.}

\item{alpha}{a numeric value for the significance level. Default is 0.1.}

\item{k.fix}{a numeric value between 0 and 1. It is the k parameter of the alternative (beta) distribution.
If it is null, it will be estimated via EM.}

\item{EM.paras}{a list of control arguments for the EM algorithm
\itemize{
\item{iterlim}{an integer value indicating the maximum number of iterations.}
\item{tol}{a numeric value giving the tolerance in the relative change in the log likelihood below which the algorithm is considered to be converged.}
\item{pi0.init, k.init}{two scalars giving the initial guess of the  pi0 and k parameter.}
\item{nlm.iter}{an integer indicating the allowed maximum iteration in running \code{'nlm'}, used to speed up computation.}
}}

\item{trace}{a logical value indicating whether to print the process.}

\item{return.model.matrix}{a logical value indicating whether to return the model matrix.  Consider setting to FALSE if it's huge.}
}
\value{
A list with the elements
\item{call}{the call made.}
\item{pi0}{a vector of the estimated null probabilities.}
\item{k}{the (estimated) value for the k parameter of the alternative distribution.}
\item{pi0.coef}{a vector of the coefficients for pi0.}
\item{gamma}{the estimated gamma value to dichotomize the p-values}
\item{loglik}{log likelihood.}
\item{EM.paras}{actually used parameters in EM algorithm.}
\item{EM.iter}{the number of iteration actually used.}
\item{rejection}{a vector containing the indices of the rejected hypotheses. }
}
\description{
The function implements a scalable, flexible, robust and powerful FWER control method for large-scale multiple testing exploiting the auxiliary covariates. 
It allows both the prior null probability to depend on covariates.
}
\examples{

data <- simulate.data(feature.no = 10000, covariate.strength = 'Moderate', covariate.model = 'pi0',
	sig.density = 'Medium', sig.strength = 'L4', cor.struct = 'None')
camt.obj.fwer <- camt.fwer(pvals = data$pvals, pi0.var = data$pi0.var)
plot.camt.fwer(camt.obj.fwer, covariate = as.vector(rank(data$pi0.var)), covariate.name = 'Covariate rank',
	log = TRUE, file.name = 'CovariateModerate.pdf')

}
\references{
Huijuan Zhou, Xianyang Zhang, Jun Chen. Covariate Adaptive Family-wise Error Control with Applications to Genome-wide Association Studies. Submitted.
}
\author{
Huijuan Zhou
}
\keyword{FWER}
