% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camt.cor.func.R
\name{plot.camt.fwer}
\alias{plot.camt.fwer}
\title{Plot the camt.fwer results}
\usage{
\method{plot}{camt.fwer}(
  camt.fwer.obj,
  covariate,
  covariate.type = c("continuous", "categorical"),
  covariate.name = "covariate",
  alpha = 0.1,
  nlimit = 10000,
  log = FALSE,
  logit = TRUE,
  file.name = "CAMT.fwer.plot"
)
}
\arguments{
\item{camt.fwer.obj}{a list returned from running 'camt.fwer'.}

\item{covariate}{a vector containing the one-dimensional covariate values.}

\item{covariate.type}{a character string of either "continuous" or "categorical".}

\item{covariate.name}{a character string for the name of the covariate to be plotted.}

\item{alpha}{a numeric value, the target FWER level.}

\item{nlimit}{an integer, the number of insignificant hypotheses to be sampled to reduce the visualization complexity.}

\item{log}{a logical value indicating whether the p-values should be plotted on the log scale.}

\item{logit}{a logical value indicating whether the pi0 should be plotted on the logit scale.}

\item{file.name}{a character string (ended with .pdf) for the name of the generated pdf file. Could include the path.}
}
\value{
A list of \code{'ggplot2'} objects.
}
\description{
The function plots the null probabilities and p-values against covariate values. Significant hypotheses are highlighted in red.
}
\examples{

data <- simulate.data(feature.no = 10000, covariate.strength = 'Moderate', covariate.model = 'pi0',
	sig.density = 'Medium', sig.strength = 'L4', cor.struct = 'None')
camt.obj.fwer <- camt.fwer(pvals = data$pvals, pi0.var = data$pi0.var)
plot.camt.fwer(camt.obj.fwer, covariate = as.vector(rank(data$pi0.var)), covariate.name = 'Covariate rank',
	log = TRUE, file.name = 'CovariateModerate.pdf')

}
\references{
Huijuan Zhou, Xianyang Zhang, Jun Chen. Covariate Adaptive Family-wise Error Control with Applications to Genome-wide Association Studies. Submitted.
}
\author{
Huijuan Zhou
}
\keyword{FWER,}
\keyword{multiple}
\keyword{testing}
