% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camt.cor.func.R
\name{plot.camt}
\alias{plot.camt}
\title{Plot the CAMT results}
\usage{
plot.camt(camt.obj, covariate, covariate.type = c("continuous",
  "categorical"), covariate.name = "covariate", alpha = 0.1,
  nlimit = 10000, log = FALSE, logit = TRUE,  file.name = "CAMT.plot")
}
\arguments{
\item{camt.obj}{a list returned from running 'camt.fdr'.}

\item{covariate}{a vector containing the one-dimensional covariate values.}

\item{covariate.type}{a character string of either 'continuous' or 'categorical'. Default 'continuous'.}

\item{covariate.name}{a character string, the name of the covariate to be plotted.}

\item{alpha}{a numeric value, the target FDR level.}

\item{nlimit}{an integer, the number of insignificant hypotheses to be sampled to reduce the visualization complexity.}

\item{log}{a logical value indicating whether the p-values should be plotted on the log scale.}

\item{logit}{a logical value indicating whether the pi0/k should be plotted on the logit scale.}

\item{file.name}{a character string for the name of the generated pdf file.}
}
\value{
A list of \code{'ggplot2'} objects.
}
\description{
The function plots the null probabilities, f1 shape parameters and p-values against covariate values. Significant hypotheses are highlighted in red.
}
\examples{

data <- simulate.data(feature.no = 5000, covariate.strength = "Moderate", covariate.model = "pi0",
	sig.density = "Medium", sig.strength = "L3", cor.struct = "None")
camt.obj <- camt.fdr(pvals = data$pvals, pi0.var = data$pi0.var, f1.var = data$f1.var, 
	alg.type = "EM", control.method = "knockoff+")
plot.camt(camt.obj, covariate = as.vector(rank(data$pi0.var)), covariate.name = "Covariate rank",
	log = TRUE, file.name = "CovariateModerate")

}
\references{
Covariate Adaptive False Discovery Rate Control with Applications to Omics-Wide Multiple Testing.
}
\author{
***
}
\keyword{FDR}
\keyword{multiple testing}
