% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camt.cor.func.R
\name{simulate.data}
\alias{simulate.data}
\title{Simulate p-values and covariates under various scenarios.}
\usage{
simulate.data(paras.mapping = paras.mapping.func(),
  covariate.strength = c("None", "Moderate", "Strong"), 
  covariate.model = c("pi0", "f1", "both"),
  covariate.dist = c('Normal', 'Uniform'),
  null.model = c('Unif', 'Left', 'Right'),
  skewness = 0.15, f1.sd = 1.0,
  feature.no = 10000, sig.dist = c('Normal', 'Gamma'), 
  sig.density = c("Low", "Medium", "High"),
  sig.strength = c("L1", "L2", "L3", "L4", "L5", "L6", "L7", "L8"),
  cor.struct = c("None", "Block", "AR1"), cor.rho = 0,
  cor.sign = c("PosCor", "PosNegCor"), cor.nblock = 500)
}
\arguments{
\item{paras.mapping}{a list with slots \code{'sig.strengths', 'sig.densities', 'pi.strengths', 'f1.strengths'}, which define the actual
numeric values used for different levels of \code{'sig.strength'}, \code{'sig.density'}, and \code{'covariate.strength'} (pi and f1).  
The mapping can be generated by invoking the \code{paras.mapping.func} function.}

\item{covariate.strength}{a character string from \code{'None', 'Moderate', 'Strong'} indicating the covariate strength.}

\item{covariate.model}{a character string from \code{'pi0', 'f1', 'both'} indicating whether the prior null probability, 
the alternative distribution or both are affected by the covariate.}

\item{covariate.dist}{a character string from \code{'Normal', 'Uniform'} indicating the distribution of the covariate.}

\item{null.model}{a character string from \code{'Unif', 'Left', 'Right'} indicating whether the null distribution of the p-value is uniform, 
left skewed or right skewed. }

\item{skewness}{a numeric value indicating the skewness of the p-value distribution under the null (mean of the z-value).}

\item{f1.sd}{a numeric value indicating the variance of the z-score under the alternative. Default is 1.}

\item{feature.no}{an integer, the number of features to be simulated.}

\item{sig.dist}{a character string from \code{'Normal', 'Gamma'} indicating the distribution of the z-value under alternative.}

\item{sig.density}{a character string from \code{'Low', 'Medium', 'High'} indicating the level of signal density.}

\item{sig.strength}{a character string from \code{'L1', 'L2', 'L3', 'L4', 'L5', 'L6', 'L7', 'L8'} indicating the level of signal strength.}

\item{cor.struct}{a character string from \code{'None', 'Block', 'AR1'} indicating the correlation structure to be simulated including
no correlation, block correlation and AR(1) correlation.}

\item{cor.rho}{a numeric value giving the correlation coefficient for \code{'Block', 'AR1'}.}

\item{cor.sign}{a character string from \code{'PosCor', 'PosNegCor'} indicating whether to simulate only positive or both positive and negative correlations.}

\item{cor.nblock}{an integer, the number of blocks to be simulated for block correlation structure}
}
\value{
A list with the elements
\item{pvals}{a numeric vector of p-values.}
\item{pi0.var}{a vector of covariate values for the prior null probability.}
\item{f1.var}{a vector of covariate values for the alternative distribution.}
\item{pi0}{ a vector of the simulated null probabilities.}
\item{truth}{a vector simulated truth for H0 (=0) or H1 (=1).}
\item{lfdr}{a vector of oracle LFDR based on the simulated pi0, f1.}
\item{fdr}{a vector of oracle FDR based on LFDR.}
}
\description{
The function simulates p-values and covariates under different signal structures (density and strength) and covariate models. 
It also allows special correlation structures among the p-values.
}
\examples{

data <- simulate.data(feature.no = 5000, covariate.strength = "Moderate", covariate.model = "pi0",
	sig.density = "Medium", sig.strength = "L3", cor.struct = "None")
camt.obj <- camt.fdr(pvals = data$pvals, pi0.var = data$pi0.var, f1.var = data$f1.var, 
	alg.type = "EM", control.method = "knockoff+")
plot.camt(camt.obj, covariate = as.vector(rank(data$pi0.var)), covariate.name = "Covariate rank",
	log = TRUE, file.name = "CovariateModerate")


}
\references{
Covariate Adaptive False Discovery Rate Control with Applications to Omics-Wide Multiple Testing.
}
\author{
***
}
\keyword{FDR}
\keyword{multiple testing}

